requirements/prod.txt
# Requirements for production

# Django
Django==1.6.5

# Better templates
MarkupSafe==0.23
Jinja2==2.7.3
jingo==0.7

# Django REST Framework
djangorestframework==2.3.14
Markdown==2.4.1
django-filter==0.7

# Django extensions
six==1.7.3
django-extensions==1.3.8

# JSON API interfaces for Django REST Framework
-e git+https://github.com/jwhitlock/drf-json-api.git@develop#egg=drf_json_api

# History of changes to models
django-simple-history==1.4.0
